pyproject.toml
[project]
name = "mx-scaling-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch==2.4.0",
    "torchvision==0.19.0",
    "torchaudio==2.4.0",
    "datasets>=2.14.4",
    "transformers>=4.46.3",
    "pytorch-ignite>=0.5.2",
]
